Production-RAG-System/requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
numpy
scikit-learn